+++ tests/models/vae_models/test_encoder.py
@@ -3,7 +3,8 @@ import torch
 from torch import nn
 
 from embkit.models.vae.encoder import Encoder
-from embkit.layers import LayerInfo, MaskedLinear
+from embkit.factory.layers import Layer
+from embkit.layers import MaskedLinear
 from embkit.constraints import NetworkConstraint
 
 
@@ -21,13 +22,13 @@ class TestEncoder(unittest.TestCase):
         self.assertEqual(tuple(z.shape), (4, 3))
 
     def test_layers_provided_require_final_equals_latent_dim(self):
-        layers = [LayerInfo(units=8), LayerInfo(units=6)]
+        layers = [Layer(units=8), Layer(units=6)]
         # final width=6 != latent_dim=4 -> should raise
         with self.assertRaises(ValueError):
             Encoder(feature_dim=5, latent_dim=4, layers=layers, batch_norm=False)
 
     def test_layers_provided_with_matching_latent_dim(self):
-        layers = [LayerInfo(units=8), LayerInfo(units=4)]
+        layers = [Layer(units=8), Layer(units=4)]
         enc = Encoder(feature_dim=5, latent_dim=4, layers=layers, batch_norm=False)
 
         # latent heads should exist
@@ -39,7 +40,7 @@ class TestEncoder(unittest.TestCase):
         self.assertEqual(tuple(z.shape), (3, 4))
 
     def test_encoder_with_masked_linear(self):
-        layers = [LayerInfo(units=7, op="masked_linear", activation=None)]
+        layers = [Layer(units=7, op="masked_linear", activation=None)]
         enc = Encoder(feature_dim=7, latent_dim=7, layers=layers, batch_norm=False)
 
         # masked layer present
@@ -58,7 +59,7 @@ class TestEncoder(unittest.TestCase):
         }
         constraint = NetworkConstraint(feature_index, latent_index, latent_membership)
 
-        layers = [LayerInfo(units=2, op="masked_linear")]  # matches latent_index length
+        layers = [Layer(units=2, op="masked_linear")]  # matches latent_index length
         enc = Encoder(feature_dim=2, latent_dim=2, layers=layers, constraint=constraint)
 
         # Should be able to call refresh_mask without error
@@ -66,7 +67,7 @@ class TestEncoder(unittest.TestCase):
         self.assertTrue(any(isinstance(m, MaskedLinear) for m in enc.net))
 
     def test_encoder_raises_on_invalid_layer_op(self):
-        layers = [LayerInfo(units=4, op="unknown_op")]
+        layers = [Layer(units=4, op="unknown_op")]
         with self.assertRaises(ValueError) as ctx:
             Encoder(feature_dim=4, latent_dim=4, layers=layers)
         self.assertIn("Unknown LayerInfo.op", str(ctx.exception))
@@ -77,7 +78,7 @@ class TestEncoder(unittest.TestCase):
         self.assertIn("latent_dim is required when no layers are provided", str(ctx.exception))
 
     def test_encoder_raises_if_latent_dim_missing_with_layers_and_latent_heads(self):
-        layers = [LayerInfo(units=4)]
+        layers = [Layer(units=4)]
         with self.assertRaises(ValueError) as ctx:
             Encoder(feature_dim=4, latent_dim=None, layers=layers)
         self.assertIn("latent_dim is required", str(ctx.exception))
@@ -111,8 +112,8 @@ class TestEncoder(unittest.TestCase):
 
     def test_encoder_layer_batch_norm(self):
         layers = [
-            LayerInfo(units=6, activation="relu", batch_norm=True),  # BN should be added
-            LayerInfo(units=2, activation=None)
+            Layer(units=6, activation="relu", batch_norm=True),  # BN should be added
+            Layer(units=2, activation=None)
         ]
         enc = Encoder(feature_dim=4, latent_dim=2, layers=layers, batch_norm=False)
